requirements.txt
pdfplumber>=0.10.0
openpyxl>=3.1.0
pandas>=2.0.0
PyQt5>=5.15.0
pytesseract>=0.3.10
pdf2image>=1.16.3
openai>=1.0.0
numpy>=1.24.0
Pillow>=10.0.0
python-dotenv>=1.0.0

